Add reward validation files and update model paths; refactor observation flattening

+++ util.py
@@ -2,6 +2,7 @@
 import os
 import torch
 from torch.autograd import Variable
+import numpy as np
 
 USE_CUDA = torch.cuda.is_available()
 FLOAT = torch.cuda.FloatTensor if USE_CUDA else torch.FloatTensor
@@ -20,7 +21,7 @@ def to_numpy(var):
 
 def to_tensor(ndarray, volatile=False, requires_grad=False, dtype=FLOAT):
     return Variable(
-        torch.from_numpy(ndarray), volatile=volatile, requires_grad=requires_grad
+        torch.from_numpy(ndarray.astype(np.float32)), volatile=volatile, requires_grad=requires_grad
     ).type(dtype)
 
 def soft_update(target, source, tau):
